Never install like this :-/
