+++ exercises/__init__.py
@@ -11,22 +11,6 @@ from os.path import (
     join as _join,
 )
 
-import pip
-
-
-def import_or_install(package, name):
-    try:
-        __import__(name)
-    except ImportError:
-        pip.main(["install", package])
-
-
-# to clone the mooc repo
-import_or_install("gitpython", "git")
-
-# to convert the notebook into a python script
-import_or_install("nbformat", "nbformat")
-import_or_install("nbconvert", "nbconvert")
 
 # constants
 _this_dir = _dirname(__file__)
