+++ utils/pip_utils.py
@@ -8,3 +8,12 @@ def get_pip_index_url() -> str:
     if pip_index_url != DEFAULT_INDEX_URL:
         dtslogger.warning(f"Using custom PIP_INDEX_URL='{pip_index_url}'.")
     return pip_index_url
+
+
+def import_or_install(package, name):
+    import pip
+
+    try:
+        __import__(name)
+    except ImportError:
+        pip.main(["install", package])
